Regular Commit

+++ croner.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
 cd ~/node.luckynode.com
-git fetch --all
-git reset --hard origin/master
+if ! git --git-dir=".git" diff --quiet
+then
+    git fetch --all
+    git reset --hard origin/master
+    forever restart nodeluckynode
+fi
